write_uploaded_file: unlink existing file rather than overwriting.
In the future, we may want to have files that are linked in multiple
places (for example, author wants to create a new version of an
existing database, and we want to implement this by hard or symbolic
links rather than duplicating every file.)

Thus, when handling a file upload, ensure that we create a new file
(using 'x'; note that python uses simply 'x' rather than 'wx') rather
than writing to any existing file.  (We do not want to read the file,
so '+' is not used.)

In case the file did already exist, try to unlink it beforehand.