get_{,main_}directory_content: generate URL for each file.
Using a single attribute, regardless of which of the three types of
URLs is used, allows simplifying logic elsewhere.

This is also likely more efficient than invoking the equivalent of the
reverse() function for every file.

Note that for the "open" case, this hardcodes the ugly "static files"
path.  Not sure what the correct way is to do this.