ComplexityValidator: do not check for too-strong passwords.
zxcvbn returns a score which is (for the current version of the
library) always an integer between 0 and 4 (lower score indicating a
"weaker" password.)  The previous code would reject passwords with a
score less than 2, or a score greater than 4.

There is no point in forbidding passwords that some future version of
the zxcvbn library considers to be "even stronger" than level 4.

Moreover, the previous code was invoking zxcvbn *twice*, which is
simply wasteful.