ComplexityValidator: devalue user/site-related terms.
If the user's password includes elements of their username, their real
name, or their email address, these elements should be treated as
having lower entropy (being more likely for an attacker to guess) than
usual.

Also, terms related to the site itself should also be treated as
having low entropy.  ('mimic' is a dictionary word, but 'physio' and
'physionet' aren't.)

This could use improvement: when an existing user is attempting to
change their password, all email addresses (past and present) should
probably be checked, not only the single "primary" address.

Note that when a new user is registering, some of the other form
fields may be missing (as, for example, in the
'test_user_creation_form' test case); in that case, the form
submission will be rejected, but we can still try to check the
password complexity at the same time.