settings.development: set ALLOWED_HOSTS to ['*'].
ALLOWED_HOSTS indicates whether Django will accept requests with a
given HTTP 'Host' header.  Setting this to '*' means requests will be
accepted regardless of the Host.  Note that manage.py still only
*listens* for connections from the local machine by default; you need
to tell it (e.g. using './manage.py runserver 0.0.0.0:8000') if you
want it to listen to other hosts.

This change simply makes it a little easier to test the development
server with different browsers in different virtual machines.