EditSubmissionForm: Add hints to the QA checklist.
The process of reviewing a project for publication is long and often
annoying.  These "hints" are not, and should not, be comprehensive;
they should remind the editor about a small number of specific issues
that are easily overlooked.

These hints are shown in the editor's console
(/console/submitted-projects/*/edit/).  The author just sees the
simple checklist.