SafeHTMLField: new class.
Having defined a whitelist of permitted HTML tags and attributes, we
need to enforce these restrictions on the server side as well as on
the client side.

This is implemented by the SafeHTMLField class, which is mostly a
drop-in replacement for ckeditor.fields.RichTextField.  When text is
submitted, it is parsed and cleaned using Bleach.  The list of
permitted tags, attributes, and style properties is derived
automatically from the 'allowedContent' option, which must be defined
in the django-ckeditor configuration (settings.CKEDITOR_CONFIGS).