physionet.css: define a default style for tables.
This is intended to be a reasonably attractive, minimal style that can
be applied to all tables.  Light gray horizontal rules (black when
printed) are displayed at the top, bottom, and between row groups.
The table is horizontally centered, and margins are added above and
below.  All cells have a modest amount of padding and are left-aligned
by default.