search.urls: rearrange project metadata URLs.
To avoid any possible confusion with actual project subdirectories,
rearrange the URLs for project metadata ('files-panel', 'get-zip', and
'view-license') by placing the version last.  This is slightly less
attractive, but these are not URLs that users will ordinarily use
directly.  (Also, none of them are actually directories.)

The 'get-file' infix is then unnecessary, so it is removed completely.

For example:
  /content/foo/1.0/files-panel/   =>  /content/foo/files-panel/1.0/
  /content/foo/1.0/view-license/  =>  /content/foo/view-license/1.0/
  /content/foo/1.0/get-zip/       =>  /content/foo/get-zip/1.0/
  /content/foo/1.0/get-file/bar   =>  /content/foo/1.0/bar