physionet_nginx.conf: make public data available via HTTP.
Public resources must be accessible via HTTP. Although it's arguably
not a great idea for clients not to support HTTPS, it's often
impossible to do so in a way that's future-proof, let alone in a way
that will both provide security for normal users, and also work
correctly in fascist-firewall environments.

And in many cases, encryption of public resources is nothing but a
needless hassle.