get_inspect_dir: don't preemptively check if path is a directory.
This function is only used by get_directory_content, to verify that
the argument is a subdirectory of the project directory.

In cases where the path doesn't exist, or isn't a directory,
list_items, and thus get_directory_content, will raise the appropriate
exception (FileNotFoundError, NotADirectoryError) which can then be
handled appropriately by the caller.