project_preview, published_project: handle missing/unreadable directory.
If the given subdirectory path doesn't exist, is not a directory, or
isn't readable, project.get_directory_content should now raise an
appropriate OSError.

In particular, if the project's toplevel directory is inaccessible for
some reason - for example, we're running a snapshot of the production
database on the staging server - we'd like the main project page to
still be readable (albeit with no list of files.)  In this case, we
return a 404 status instead of the usual 200, and display an error
message similar to the message displayed if the user doesn't have
permission to view the project files.