Import LightWAVE client version 0.68.
LightWAVE is tracked in a separate repository, which is currently
hosted at <https://github.com/bemoody/lightwave>.

The static files here are copied directly from the standard LightWAVE
client (which itself includes copies of the jquery, jquery-ui, and
jquery-cookie libraries.)

The HTML template (home.html) is a copy of lightwave.html.  Because
LightWAVE is meant to be usable as a standalone application, this page
is a complete HTML document rather than being derived from base.html.
Additional headers and footers, to match the site style, are defined
in lw-head.html, lw-top.html, and lw-bottom.html.