Add LightWAVE URL handlers.
Rendering the main page (/lightwave/, /lightwave/projects/<NAME>/) is
straightforward.  The server URL is set appropriately (using the new
'default_server' feature of LightWAVE 0.68.)  The scribe URL is set to
point to archive.physionet.org for the time being; this will need to
be fixed.

Static files (js/, css/, doc/), which are referenced by relative URLs
from the main page, are handled by redirecting to the corresponding
Django static resource path.  This is a bit ugly but should work.

The LightWAVE server is a separate program (written in C and using the
WFDB library) which must be installed separately; the serve_lightwave
function invokes this server executable to do its work.  For
public-facing servers, it is strongly recommended to use the
*sandboxed* variant to protect against malicious uploaded files.  The
sandboxed server requires libseccomp and Linux 3.5 or later, and must
be installed setuid root.