Enforce a strict ordering of project authors.
This will add a database constraint, which could conceivably fail if
any existing projects were created incorrectly, so take care and test
before applying to the live system.

Swapping the order of authors was previously done using two updates,
which would of course temporarily violate this constraint.  So instead
we now do three updates (move Alice to position 0, then move Bob to
Alice's former position, then move Alice to Bob's former position)
wrapped in an atomic transaction.

It seems like it ought to be possible to do the swap as a single
update, but the obvious approach didn't seem to work.