physionet.css: better rules for "sticky" <footer> and padding <main>.
If the "flexbody" class is applied to the body element, these rules will:
- ensure that the body is at least as tall as the browser window
- add any unused vertical space to the <main> element

with the effect that, if the page is small enough to fit in the
window, the footer element is placed at the bottom.

(For the detailed reasoning behind these rules and how they work
around bugs in some older browsers, see:
https://philipwalton.com/articles/normalizing-cross-browser-flexbox-bugs/)

In addition, vertical margins are added to <main>, instead of
overloading the "container" class.