News: add guid field.
This field must contain a universally unique identifier (not
necessarily a UUID, just a string that is universally unique), which
can be used by news readers to identify the story.

By defining this field with 'default=uuid.uuid4', a random UUID will
be generated when a new news item is created, and also for each
existing news item when the migration is applied.

(Note that it's not possible to make 'guid' the primary key, or add a
uniqueness constraint, while performing an automatic migration,
because existing items don't have a guid yet.)