Add MathML to the default allowedContent.
This whitelist permits most presentational MathML tags and attributes.
It also permits some classes that are not part of MathML, but are used
in the MathML that is *generated* by MathJax, and are used in turn to
correctly render some constructs, such as \mathit{}.

The use of classes may be somewhat confusing here, as CKEditor's
content filter is somewhat different, and more fine-grained, than
Bleach.  With the rules given here:

- CKEditor will permit classes 'MJX-TeXAtom-ORD', 'MJX-TeXAtom-OP',
  etc., for any element (although they won't have any effect except on
  certain MathML elements.)

- Bleach will permit any classes, but only for <mi>, <mn>, <mo>,
  <mrow>, and <mtext> elements.

This whitelist may be incomplete.