pnmathml: store math in document as MathML.
Newly inserted equations will be converted into MathML and the
original TeX syntax included as an annotation.  For example, the
default 'x^2' will be stored in the document as

<math>
  <semantics>
    <msup>
      <mi>x</mi>
      <mn>2</mn>
    </msup>
    <annotation encoding="application/x-tex">x^2</annotation>
  </semantics>
</math>

The <annotation> tag makes it possible to edit the equation later
(e.g. by double-clicking), and can also be viewed in the final page
using the MathJax context menu.