pnmathml: add blockEquation widget.
The existing pnmathml widget is renamed to 'inlineEquation'.

Block equations are represented by a 'math' element with the 'display'
attribute set to 'block'.  This causes the equation to be shown in
"display style" (centered on a line of its own, full-size fractions,
over/underscripts for summations, etc.)  The equation is also wrapped
in a 'div' element for the benefit of HTML parsers that don't
understand MathML.

The widget itself behaves much like the inlineEquation widget, except
that it is treated as a block element rather than inline, and the
equation is rendered using \[...\] instead of \(...\).