pnmathml: align inline widgets with the text baseline.
In order to determine the position of the inner text baseline
(relative to the borders of the iframe), we insert an invisible span
element adjacent to the preview element.  MathJax will ensure that the
preview element's baseline is correctly aligned with the invisible
span, and since the invisible span has zero descent, this lets us
calculate the overall ascent and descent of the iframe contents.

We don't do this in the "block equation" case because the invisible
element would force MathJax to add extra vertical padding in that
case.