pnmathml: use original MathML instead of TeX annotation.
If there is a discrepancy between the MathML and TeX in the underlying
document (for example, because it has been edited by hand, or because
it was generated using a different version of MathJax), the WYSIWYG
view should display the equation as it will appear in the final
document.

Double-clicking the equation will allow viewing/editing the TeX
(annotation) source, and clicking "OK" will then cause the MathML to
be regenerated.

For this to work properly, mathJaxLib must point to a configuration
that supports both TeX and MathML input, such as
TeX-AMS-MML_HTMLorMML.