pnmathml: add requiredContent.
This indicates what content the widget "requires" in order to be
functional; if the given content is not permitted by the
allowedContent rules, the widget buttons will be hidden.

The inlineEquation widget requires the math tag, and the blockEquation
widget requires the math tag and display attribute.  (Both widgets, in
fact, require substantially more than this, but 'requiredContent' can
only contain a single rule.)