ckeditor config: enable "native" spell checking.
This means the web browser's built-in spell checker will be used if
possible.

(In either Firefox or Chromium, you can normally right-click on a
misspelled word to view suggestions.  This doesn't exactly work with
CKEditor, unfortunately, as CKEditor grabs the mouse events for
itself.  It may be necessary to hold Shift or Control while clicking
in order to open the browser's context menu.  This is not ideal, but
seems better than nothing.)