pnmathml: update preview on input event rather than keyup.
We want to update the preview when the text in the input field
changes, and only then.  Keys can be pressed (such as arrow keys)
without modifying the text, and the text can be modified without
pressing any keys (such as copying and pasting using the mouse.)

Presumably 'keyup' was originally used because back in the day, many
browsers didn't support the 'input' event (they did support the
'change' event, but that event typically doesn't fire until the user
moves focus *out* of the input field.)

All current browsers, and most of the obsolete browsers still in use,
support the 'input' event.  Internet Explorer 9's implementation of
'input' is buggy, but a simple workaround is included here.  (On older
versions of IE, many other site features are broken or degraded; IE 9
is the oldest version that mostly works.)