Cleaner style for directory breadcrumbs in file panels.
Remove various NBSPs used for padding, and remove inter-tag spaces so
that the text of the breadcrumbs (e.g. if copied and pasted) can be
read as a filesystem path.

In place of spaces and NBSPs, use CSS to add a quarter em of padding
around each slash, to make the path easier to read.