get_project_file_warning: new function.
This function is meant to be called after get_project_file_info(), to
check for potentially problematic filenames and display a warning
message.

In addition to checking for individual filenames that fail
validate_filename(), it checks for pairs of files that differ only in
case.  Some popular operating systems brokenly fail to handle such
files.