ActiveProjectFilesForm: validate subdir parameter.
This form parameter is used to construct the working directory
(file_dir) used for other file operations.  The 'subdir' value is
concatenated onto the project storage directory using os.path.join.

Thus, setting 'subdir' to either an absolute path, or a path with
'..', allows any authenticated user to (blindly) create, delete,
rename, or move any files or directories, with the permissions of the
django server.

This applies to all five of the actions that can be triggered by
project_files, and requires only that the user have an account and
create a project.