test_project_files: check that invalid subdirs are rejected.
CreateFolderForm, EditItemsForm (i.e. delete), and UploadFilesForm are
tested to be sure that these forms will correctly reject invalid
'subdir' values.

RenameItemForm and MoveItemsForm currently are *not* tested, because
both of those will currently crash if subdir is invalid.  Fixing those
bugs is outside the scope of the present security fix.