project.utility.rename_file: new function.
This function renames a file or directory (as os.rename does), but
attempts to avoid overwriting any existing file.  A FileExistsError
will be raised if a conflict is detected.

Note that rename() will fail if:
 - the old path is a directory and the new path is an existing
   non-directory
 - the old path is a directory and the new path is an existing,
   non-empty directory
 - the old path is a non-directory and the new path is an existing
   directory

but it will clobber the destination if:
 - the old path is a directory and the new path is an existing empty
   directory
 - the old path is a non-directory and the new path is an existing
   non-directory

Another, more complicated option would be to use link() followed by
unlink() for non-directories, and to use mkdir() followed by rename()
for directories; but this still would not avoid all possible race
conditions, and still would require at least two system calls for
every file.

Recent versions of Linux provide renameat2(), but Python doesn't yet
include that function.