EditItemsForm: don't check filenames against directory list.
For all three of these forms (DeleteItemsForm, RenameItemForm, and
MoveItemsForm), the previous code would read the contents of the
directory and then check whether each supplied "item" was in the list.

Instead, simply check whether the given arguments are potential
filenames (validate_oldfilename); the clean and/or perform_action
functions will now handle missing files as they do other file I/O
errors.

In particular, for DeleteItemsForm and MoveItemsForm, this means that
if some of the selected files no longer exist, the *other* selected
files will still be deleted/moved (and an error message will be shown
about the files that don't exist.)

In RenameItemForm, keep 'items' as a list for consistency (rather than
having a field called 'items' whose value is a single item), and raise
a validation error unless exactly one item is selected.