remove_items: add optional ignore_missing argument.
If this argument is false, raise a FileNotFoundError if one of the
given paths does not exist.  Otherwise, non-existent paths are
silently ignored as before.

In addition, try to unlink the file and catch the IsADirectoryError,
rather than first statting the file to decide whether to call
os.remove or shutil.rmtree.

Note that this means the function now *can* remove:
 - symbolic links to directories
 - broken links, fifos, devices, sockets

Previously, trying to remove a symbolic link to a directory would
raise OSError("Cannot call rmtree on a symbolic link"), and trying to
remove a non-regular, non-directory file would be silently ignored.