MoveItemsForm: require subdirectories to be supplied by caller.
All views that render a MoveItemsForm also invoke
get_project_file_info, which reads the current directory ("subdir")
and generates sorted lists of files and subdirectories for display.

Thus, we don't need to read and sort the directory a second time just
to generate a list of possible target directories - simply require the
caller to pass the display_dirs list to get_file_forms.