utility.serve_file: add optional allow_directory argument.
Normally (when running under nginx), this function simply converts the
given file path into a virtual URL which is served by nginx using the
X-Accel-Redirect mechanism.  This in turn means that:

- if the path doesn't end with a slash and it refers to a regular
  file, nginx will serve the contents of the file

- if the path ends with a slash and refers to a directory (and
  directory listings are permitted in that location), nginx will serve
  a directory listing

- if the path refers to a directory but doesn't end with a slash,
  nginx will redirect the client accordingly

When running in development mode, we need to emulate this behavior (so
we can test, e.g., that recursive wget works correctly).  Don't do
this by default (so we don't inadvertently permit directory listings
somewhere we don't want to), but do so if 'allow_directory' is set.