custom-mime.types: set default type to text/plain.
Although this is often completely wrong, it seems to give better
results than not giving a content type at all. :/

For example, a python source file called 'foo.py' should be displayed
as plain text, but many browsers will, by default, treat this as
'text/x-python' and insist that the user download the file or open it
in an external program.

On the other hand, if the file actually *isn't* text but is served
with content type text/plain, the browser will usually recognize that
it isn't text and refuse to display it.