utility.serve_file: set explicit content type in development mode.
When running under nginx, we set 'Content-Type' to an empty string and
let nginx assign an appropriate content type to the file.  (If we
don't set Content-Type at all, it is set to 'text/html' by default,
which we really don't want.)

When running in development mode, we want to ensure that 'text/plain'
is used as the default content type (otherwise the browser will make
bad assumptions.)

The current set of content types defined by file_content_type is
incomplete and is somewhat inconsistent with nginx's (and
WSGIServer's) behavior, but should be better than nothing.