short_description: allow any text.
Allow any text (up to 250 characters) in the short_description field,
instead of restricting it to "validate_alphaplusplus"
([\w][\'\,\.\w\ -]*).

This applies to ActiveProject, ArchivedProject, and PublishedProject.

(The only place these fields are currently used, apart from the forms
that edit them, is in search/content_list.html, where it is properly
escaped.)

While we are performing a migration (which should be a no-op), also
remove the unnecessary default value.