*_form_snippet.html: use field.html_name instead of field.name.
If a form is created with a prefix (as in the case of the new
credentialing form, which uses the prefix 'application'), the actual
form field names used in HTML are not the same as the 'name' of the
field.

'field.html_name' appears to be the correct thing to use instead of
'field.name' when referring to the field on the HTML/DOM side (e.g.,
in <label for="XYZ">.)  If the form has no prefix, the two should be
equivalent.

https://stackoverflow.com/questions/18123695/