zip_file: use the system zip command instead of zipfile.
The Info-ZIP zip command permits the use of multiple zlib compression
levels (-9 being the highest), and in cases where the compressed file
is larger than the uncompressed file, will automatically store the
file uncompressed.

(In contrast, the Python zipfile library requires picking a
compression method ahead of time, and always uses the corresponding
compression library's default settings.)

We use the '-X' option to zip to avoid including additional metadata
that is generally useless and irreproducible.  We still use
sorted_tree_files, and pass the list of files via standard input, so
that the order of files in the archive is predictable.

In addition, the output archive is initially written to a temporary
file and only moved to the destination filename when it is complete.