physionet_nginx.conf: fix behavior of some old URLs.
Top-level directories under /physiotools/ and /physiobank/database/
are handled by django, which will return either a 404 or a 302
depending on whether or not there is a project with that name.

  /physiobank/database/mitdb/ (=> /content/mitdb/)
  /physiobank/database/fnord/ (404)
  /physiotools/ecgpuwave/ (=> /content/ecgpuwave/)
  /physiotools/fnord/ (404)

However, if the directory contains an 'index.html', that file is
served instead.

  /physiotools/wag/
  /physiotools/wpg/
  /physiotools/wug/

Existing files are served as-is.

  /physiobank/database/wfdbcal
  /physiobank/database/mitdb/100.dat
  /physiotools/archives/wfdb-10.6/wfdb-10.6.2.tar.gz
  /physiotools/ecgpuwave/src/ecgpuwave-1.3.4.tar.gz

Existing non-top-level subdirectories are served as a directory
index (or using index.html if that exists.)

  /physiotools/debian/dists/jessie-challenge-2017/

Non-existent files within /physiotools/ are handled by django which
will typically show a 404 page.

Non-existent files within /physiobank/database/ are typically
requested by WFDB clients and thus are *not* handled by django (for
performance, and also because many such clients don't support HTTPS
nor do they understand redirection.)  These will show nginx's basic
404 page.