requirements.txt: add package hashes.
Add SHA-256 hashes for all packages, so that pip will verify that it
is installing what was expected.

Note that if a hash is given for *any* package in requirements.txt,
then a hash must be given for *every* package.  This also means that
all indirect requirements will now need to be listed explicitly.

The list is now divided into direct and indirect requirements; note
that:

* sqlparse and django-js-asset do not appear to be used directly (but
  are required by Django and django-ckeditor), so these have been
  demoted to indirect requirements.

* chardet wasn't explicitly listed before (although it's required by
  requests (yes, really)), but we do use it in project.fileviews, so
  it has been promoted to a direct requirement.

For a few of these packages, pip wants to install a binary package.
In these cases, the first hash listed is for Python 3.5 (Debian 9) and
the second is for Python 3.7 (Debian 10).