Added a way to save the user cloud information to the user profile
Here I do several things,
1. There is a new credentials file for GCP. This file will handle
the delegation of permissions to create groups in the google
PhysioNet organization
2. There are new environment variables that are required for the
AWS Open Data platform. Those variables are used to automatize the
deploying to the cloud
3. Adding two new models, one to store the user cloud information,
and the second to store the cloud file information to then display
in the published project page.
4. Moved the GCP credentials to the base settings file because it
doesn’t change in between develop and production.
5. Show different ways to download the project files