Fixing second user bug
On registration the username was being forced to lower case.
I removed that, and changed the defined user get queries to
be case insensite.

Now admin is the same as Admin.

The thing to note here is that for future code, we need to
use "User.objects.get(username__iexact="
instead of "User.objects.get(username="