Setting a proper except for the pagination
Here is a problem is I do a pop.
If I simply do a pop, and the item exists it will return an array.
The function “paginator.page” takes in a number and not an array.

There are so many ways to do this then, I would do:
page = params.pop('page', 1)
if 'page' in params:
    page = page[0]

page = params.get('page', 1)
if 'page' in params:
    params.pop('page')

or just ignore the whole params variable and keep it how it was in the beginning.