Moved pagination to another file and cleaned past credentialed applications
Here remove the try except and use the built ‘get_page’ in as suggested.
I moved it to another file, and just call the function instead of having the same thing in multiple places.

Also, I cleaned the past credential applications page. Now we don’t use 2 variables to differentiate legacy and new applications.

I was going to add pagination here, but it’s a bit more complicated because of the tab. To add pagination, the pages has to be separated.