Disable autoescape for credentialing verification messages.
When someone applies for credentialing, these templates are used to
construct a message that is sent to the person listed as a reference.

This text isn't HTML, so if the applicant's research summary includes
<, >, &, ", or ', those characters should not be escaped.

This change also makes these templates consistent with other templates
that are used to generate email messages.