notification.utility.get_url_prefix: new function.
This function returns a string consisting of protocol, hostname, and
port number if necessary, with no trailing slash.  This information is
obtained from the supplied HttpRequest object.

In normal deployment, the protocol should be 'https', but for testing
the server locally we want 'http' instead.  Likewise, in normal
deployment the hostname is derived from django.contrib.sites, but in
testing, we use the hostname and port from the HTTP request.