Use {{ url_prefix }} in place of http://{{ domain }}.
In numerous email templates, we normally want to use an 'https' URL
rather than 'http'.  However, for ease of testing, we still want to
use 'http' when running a (local, insecure) development server.

This is done using a new template parameter 'url_prefix', which
includes both protocol and domain.  The protocol is set to 'https' if
the original request was made using HTTPS (which should always be the
case in production.)