reset_password_email: always use 'https' as the protocol.
The actual rendering of this template is controlled by
django.contrib.auth.views.PasswordResetView, which has some
complicated logic of its own to determine whether {{ protocol }}
should be 'http' or 'https'.

I can't be bothered to untangle said logic at the moment, so I'm
hardcoding this particular template to https.