Check if the add user to the google group doesnt fail
Since google doesnt handle dots in the email if a person
writes person@gmail.com is the same as p.erson@gmail.com

Since I cannot check if the email exists in the google group,
then I have to do an exact match, but if the person does the
stated above then I will get a 500 error because a 500 error page.

Here I check if the person exists in the group, and if it doesnt
then add the person. If the person does exists then check the error.