send_contact_message: use noreply@ as mail sender.
The name and address supplied by the user are shown in the 'From'
header, but the 'Sender' header and envelope sender are set to
settings.DEFAULT_FROM_EMAIL (noreply@physionet.org).  This should
ensure that bounce messages are returned to us (not to some other
innocent mail server), and avoid problems due to SPF and similar.