Define SaferAdminEmailHandler class.
This class is a slightly-altered version of the built-in
AdminEmailHandler class (which sends error messages by email to the
administrator); it omits most potentially-private information.

In particular, the built-in error message includes the values of all
POST variables (unless manually suppressed), and the values of all
cookies, UWSGI variables, and HTTP request headers.  All of these may
contain private information submitted by the client.  Here, all of
this information is replaced with stars.

The built-in error message also includes all attributes of the
settings module (unless they meet certain hard-coded ad-hoc rules for
exclusion), which may contain server private data, and generally just
serves to add clutter.  This information is omitted entirely.

As described in <https://stackoverflow.com/a/36383577>, there doesn't
seem to be any way of customizing the format of email error reports
other than by subclassing AdminEmailHandler and including some
boilerplate code.