console.tasks: new module.
This module answers the question:

  "Given a model instance X, are there any background tasks that are
  scheduled to operate on X?"

In particular, for long-running operations that can potentially break
things if performed in the wrong order, we would like the
administrators to be able to see at a glance what tasks are pending
for a particular PublishedProject.

This module works by keeping track of which background task functions
have parameters that refer to an object.  (This is done by adding the
'associated_task' decorator to each background task function.)

Then, if we want to know what tasks are pending on a particular object
X, we look through the existing Task objects, and parse their
JSON-encoded parameters to find which ones refer to X.

This is somewhat inefficient, but should be adequate for purposes of
managing adminstration tasks.  It keeps things simple at the API level
and avoids delving too deeply into the innards of the background_tasks
package.