manage_published_project: show pending tasks.
On the published project management page
(/console/published-projects/demobsn/1.0/), there are various buttons
that the administrator can press to trigger asynchronous actions.

To help avoid accidentally performing these tasks in the wrong order,
or performing the same task twice, display a list of pending tasks on
the management page, and if there are any tasks pending, disable the
buttons for "conflicting" tasks.

("Make Checksum File" and "Make Zip" are "read-write" tasks, and no
other tasks should be scheduled for the same project at the same time.
"Create GCP bucket and send files" is a "read-only" task, that in
principle can be performed in parallel with other tasks - currently
there aren't any others.)

This doesn't prevent the administrator from editing the page source
and submitting the tasks anyway.