move_files_as_readonly: remove confusing no-op conditional.
move_files_as_readonly always invokes make_special_files.  There is
(currently) no such thing as a project that has no files anyway.

The way this function was previously written, it was apparently meant
to skip calling 'make_special_files' if the project directory was
empty.  This was not actually the case: the "if" statement was simply
checking whether 'published_project.storage_used' was a true value.
Since 'published_project.storage_used' is a function (not a number),
it was always true.