StorageInfo: permit used to be None.
This allows a StorageInfo object to represent a project where the
total file size is unknown.  If 'used' is None, the "storage used" and
"storage remaining" will be displayed as "unknown".

Moreover, simplify the logic by ignoring the 'include_remaining'
argument.  The 'remaining', 'readable_remaining', 'p_used', and
'p_remaining' attributes are always defined, regardless of
'include_remaining'.  There is no point in complicating the logic here
just to avoid a bit of trivial arithmetic.