get_storage_info: add optional force_calculate argument.
In general, we do not ever want the server to recursively scan an
entire project directory in order to serve a single web page.  A
project directory may recursively include millions of files, in some
cases.  Simply performing the equivalent of 'du' may take
entire *minutes* if the directories and inodes aren't in cache, which
is far too long to wait.

In the case of published projects, the total size is calculated
(multiple times, in fact) and stored in the model's main_storage_size
field.  So we can use this information and avoid scanning the
directory at all.  (In fact, the only place where this appears to be
used is in console.views.manage_published_project.)

For active projects, this problem will eventually be solved by the use
of disk quotas, but for the time being the only way to know how much
space is used is to add it all up.  So for the time being, give the
caller two options:

- if force_calculate is true (default), invoke storage_used as before,
  which may take a very long time;

- if force_calculate is false, return a dummy StorageInfo object that
  states the usage is unknown.

Once we have working disk quotas, this will be fixed and the
force_calculate argument will probably go away.