info_card: add optional force_calculate argument.
If this argument is false (default), then do not do any expensive
operations to calculate the project file size.  This means that in the
various console pages related to an active project:

  /console/submitted-projects/*/info/
  /console/submitted-projects/*/edit/
  /console/submitted-projects/*/copyedit/
  /console/submitted-projects/*/awaiting-authors/
  /console/submitted-projects/*/publish/

the size will be displayed as "unknown".

Currently, 'info_card' is only used by the above pages and by

  /console/published-projects/*/*/

(but in the case of published projects, the size is stored in the
model so does not need to be calculated.)

Setting 'force_calculate' to True is not recommended but it may be
necessary in some cases.  Eventually we will have working disk quotas
and all of this will go away.