Update AnonymousAccess model
Project was altered to be a GenericFK in order to
allow anonymous access to both published and active projects.

Passphrase is a hash from a random string of length 20
and is set by the function `set_passphrase`, which returns the raw key.
Once the raw key has been set it cannot be recovered, as it is not stored.

Functions to generate and check a passphrase was added to Metadata,
and the function `has_access` has been updated and tries to validate
a anonymous access if the argument passed is a string.