messages, fix authentication, and limiting access generation
Provide messages to users accessing resources through
anonymous access to inform that content is not static and provide
the last modification times. Also inform the authors that changes
will affect anyone acessing through this feature.

Allow anonymous access to lightwave and fix the problem that
cookies generated by the anonymous login were not project specific.

This also limits the generation of anonymous access tokens to
projects that have been accepted by the editor.