Add VerboseStreamHandler to logging configuration.
This should cause detailed messages to be printed to standard error
(when testing) or to the uwsgi log file (on live server) when a
request hits a fatal exception.

In development, we don't need to display two copies of the same
information, so disable the mail_admins handler if debug mode is
enabled.