Fixes the test error in the console
There was a 40 != 25 error in a test of the console. The error I received was while trying to delete the file, the except returned a “PermissionError” instead of a “IsADirectoryError”.

Here I now check if it’s a link because of versions, and then unlink.
OR, if it’s a folder, and delete the folder with its content.
Last if it’s a file, and then delete it.