Fixed PR as recommended
I like the option “#messy” because it differentiates between permission denied and operation not permitted.

Since the errors can be between is a directory or Operation not permitted depending on the OS.
Here we catch them both. If the error is not one of those, then it raises the error.

Note, this is different from permission denied.